+++ monument/cli/src/lib.rs
@@ -12,8 +12,7 @@ pub mod spec;
 use std::{
     fmt::Write,
     io::Write as IoWrite,
-    num::ParseIntError,
-    path::{Path, PathBuf},
+    path::Path,
     str::FromStr,
     sync::{
         atomic::{AtomicBool, Ordering},
@@ -22,10 +21,6 @@ use std::{
     time::{Duration, Instant},
 };
 
-use bellframe::{
-    place_not::{self, PnBlockParseError},
-    InvalidRowError,
-};
 use log::{log_enabled, LevelFilter};
 use monument::{Comp, Config, Progress, Query, QueryUpdate};
 use ordered_float::OrderedFloat;
@@ -47,7 +42,7 @@ pub fn run(
     debug_option: Option<DebugOption>,
     config: &Config,
     ctrl_c_behaviour: CtrlCBehaviour,
-) -> Result<Option<QueryResult>, Error> {
+) -> anyhow::Result<Option<QueryResult>> {
     /// If the user specifies a [`DebugPrint`] flag with e.g. `-D layout`, then debug print the
     /// corresponding value and exit.
     macro_rules! debug_print {
@@ -62,8 +57,7 @@ pub fn run(
     let start_time = Instant::now();
 
     // Generate & debug print the TOML file specifying the search
-    let spec =
-        Spec::read_from_file(input_file).map_err(|e| Error::SpecFile(input_file.to_owned(), e))?;
+    let spec = Spec::read_from_file(input_file)?;
     debug_print!(Spec, spec);
 
     // Convert the `Spec` into a `Layout` and other data required for running a search
@@ -73,7 +67,9 @@ pub fn run(
     debug_print!(Layout, &query.layout);
 
     // Optimise the graph(s)
-    let graph = query.unoptimised_graph(config).map_err(Error::GraphGen)?;
+    let graph = query
+        .unoptimised_graph(config)
+        .map_err(|e| anyhow::Error::msg(graph_build_err_msg(e)))?;
     debug_print!(Graph, graph);
     let optimised_graphs = query.optimise_graph(graph, config);
     if debug_option == Some(DebugOption::StopBeforeSearch) {
@@ -155,33 +151,6 @@ impl QueryResult {
     }
 }
 
-/// The possible ways that a run of Monument could fail
-#[derive(Debug)]
-pub enum Error {
-    SpecFile(PathBuf, spec::TomlReadError),
-    MusicFile(PathBuf, spec::TomlReadError),
-
-    PartHeadParse(InvalidRowError),
-    // TODO: For too short inputs, suggest expanding with either bells or a `*`
-    ChMaskParse(String, bellframe::mask::ParseError),
-    ChPatternParse(String, bellframe::mask::ParseError),
-
-    NoMethods,
-    CcLibNotFound,
-    MethodNotFound {
-        suggestions: Vec<String>,
-    },
-    MethodPnParse(PnBlockParseError),
-
-    CallPnParse(String, place_not::ParseError),
-    /// The user set both `bobs_only = true` and `singles_only = true`.
-    BobsOnlyAndSinglesOnly,
-    LeadLocationIndex(String, ParseIntError),
-
-    LayoutGen(monument::layout::new::Error),
-    GraphGen(monument::graph::BuildError),
-}
-
 /// What item should be debug printed
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
 pub enum DebugOption {
@@ -210,6 +179,20 @@ impl FromStr for DebugOption {
                 v
             )),
         })
+    }
+}
+
+////////////////////
+// ERROR MESSAGES //
+////////////////////
+
+fn graph_build_err_msg(e: monument::graph::BuildError) -> String {
+    use monument::graph::BuildError as BE;
+    match e {
+        BE::SizeLimit(limit) => format!(
+            "Graph size limit of {} nodes reached.  You can set it higher with `--graph-size-limit <n>`.",
+            limit
+        )
     }
 }