+++ monument/lib/src/layout/new/mod.rs
@@ -1,5 +1,7 @@
 //! Utility code for building [`Layout`]s in common scenarios.
 
+use std::fmt::{Display, Formatter};
+
 use bellframe::{method::RowAnnot, Bell, Block, Mask, PlaceNot, Row, Stage};
 use itertools::Itertools;
 use serde::Deserialize;
@@ -38,7 +40,10 @@ impl Layout {
 #[derive(Debug, Clone)]
 pub enum Error {
     NoMethods,
-    UndefinedLeadLocation(String),
+    UndefinedLeadLocation {
+        call_name: String,
+        label: String,
+    },
     DuplicateShorthand {
         shorthand: String,
         title1: String,
@@ -62,6 +67,62 @@ pub enum Error {
         input_mask2: Mask,
     },
 }
+
+impl Display for Error {
+    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
+        match self {
+            Self::NoMethods => write!(f, "Can't have a composition with no methods"),
+            Self::DuplicateShorthand {
+                shorthand,
+                title1,
+                title2,
+            } => write!(
+                f,
+                "Methods {:?} and {:?} share a shorthand ({})",
+                title1, title2, shorthand
+            ),
+            Self::UndefinedLeadLocation { call_name, label } => write!(
+                f,
+                "Call {:?} refers to a lead location {:?}, which doesn't exist",
+                call_name, label
+            ), // TODO: Suggest one that does exist
+            Self::CallingPositionsTooShort {
+                call_name,
+                calling_position_len,
+                stage,
+            } => {
+                write!(
+                    f,
+                    "Call {:?} only specifies {} calling positions, but the stage is {}",
+                    call_name, calling_position_len, stage
+                )
+            }
+            // TODO: Rename 'calling bells' to 'observation bell'
+            Self::ConflictingCallingBell((mask1, calling_bell1), (mask2, calling_bell2)) => {
+                write!(
+                    f,
+                    "Conflicting observation bells:\n      {} wants {}\n  but {} wants {}",
+                    mask1, calling_bell1, mask2, calling_bell2
+                )
+            } // TODO: Make a test case for this once custom calling bells are possible
+            Self::AmbiguousCourseHeadPosition {
+                mask1,
+                input_mask1,
+                mask2,
+                input_mask2,
+            } => {
+                write!(
+                    f,
+                    "The same course could be given two different course heads:"
+                )?;
+                write!(f, "\n     {}, satisfying {}", mask1, input_mask1)?;
+                write!(f, "\n  or {}, satisfying {}", mask2, input_mask2)
+            }
+        }
+    }
+}
+
+impl std::error::Error for Error {}
 
 pub type Result<T> = std::result::Result<T, Error>;